W-16171426: [Android] Update Native Login Sample App to include QR Code Login Flow (Documentation Comments To Enable Log In Via UI Bridge API Generated QR Codes)
