+++ AndroidNativeKotlinTemplate/app/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 
 dependencies {
     implementation("com.salesforce.mobilesdk:MobileSync:12.2.0")
+    // Uncomment when enabling log in via Salesforce UI Bridge API generated QR codes
     implementation("com.google.zxing:core:3.4.1")
     implementation("com.journeyapps:zxing-android-embedded:4.3.0")
 }
