+++ AndroidNativeKotlinTemplate/app/src/main/AndroidManifest.xml
@@ -15,7 +15,6 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:label="@string/app_name"
             android:theme="@style/SalesforceSDK">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -24,12 +23,14 @@
         </activity>
 
         <!-- Login activity -->
-        <activity android:name="com.salesforce.androidnativekotlintemplate.LoginActivity"
+        <!-- Use `android:name="com.salesforce.androidnativekotlintemplate.QrCodeEnabledLoginActivity"` when enabling log in via Salesforce UI Bridge API generated QR codes -->
+        <activity
+            android:name="com.salesforce.androidnativekotlintemplate.QrCodeEnabledLoginActivity"
             android:theme="@style/SalesforceSDK"
             android:launchMode="singleTask"
             android:exported="true">
 
-            <!-- Login with QR -->
+            <!-- Uncomment when enabling log in via Salesforce UI Bridge API generated QR codes -->
             <intent-filter>
                 <data
                     android:host="android"
