+++ AndroidNativeKotlinTemplate/app/src/main/java/com/salesforce/androidnativekotlintemplate/MainApplication.kt
@@ -40,7 +40,16 @@ class MainApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        MobileSyncSDKManager.initNative(applicationContext, MainActivity::class.java, LoginActivity::class.java)
+        MobileSyncSDKManager.initNative(
+            applicationContext,
+            MainActivity::class.java,
+            /*
+             * Enable login via Salesforce UI Bridge API generated QR code using
+             * a custom log in activity.
+             *
+             */
+            // Uncomment when enabling log in via Salesforce UI Bridge API generated QR codes
+            QrCodeEnabledLoginActivity::class.java)
         MobileSyncSDKManager.getInstance().registerUsedAppFeature(FEATURE_APP_USES_KOTLIN)
 
         /*
@@ -58,6 +67,8 @@ class MainApplication : Application() {
 		 */
         // MobileSyncSDKManager.getInstance().pushNotificationReceiver = pnInterface
 
-        MobileSyncSDKManager.getInstance().isQRLoginFlowEnabled = true
+        // Enable login via Salesforce UI Bridge API generated QR code.
+        // Uncomment when enabling log in via Salesforce UI Bridge API generated QR codes
+        MobileSyncSDKManager.getInstance().isQrCodeLoginEnabled = true
     }
 }
