+++ AndroidNativeKotlinTemplate/app/src/main/res/values/strings.xml
@@ -9,6 +9,8 @@
     <string name="fetch_accounts_button">Fetch Accounts</string>
     <string name="clear_button">Clear</string>
 
+    <!-- Text for the "Log In With QR" button -->
+    <!-- Uncomment when enabling log in via Salesforce UI Bridge API generated QR codes -->
     <string name="login_with_qr">Log In with QR</string>
 
 </resources>
